requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.11
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
